+++ tower/CHANGELOG.md
@@ -6,20 +6,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
+None.
+
+# 0.4.5 (February 10, 2021)
+
+### Added
+
 - **util**: Add `ServiceExt::map_future`. ([#542])
-- **util**: Make combinators implement `Debug` in more cases
-- `tracing` is now only pulled in for the features that need it.
-- **util**: Add `option_layer` to convert an `Option<Layer>` into a `Layer`. ([#555])
 - **builder**: Add `ServiceBuilder::option_layer` to optionally add a layer. ([#555])
 - **spawn-ready**: SpawnReady now propagates the current `tracing` span to
   spawned tasks ([#557])
-- **steer**: `Steer` now implements `Clone`.
 - **make**: Added `Shared` which lets you implement `MakeService` by cloning a
-  service.
+  service. ([#533])
+
+### Fixed
+
+- **util**: Make combinators implement `Debug` in more cases. ([#552])
+- `tracing` is now only pulled in for the features that need it. ([#551])
+- **steer**: `Steer` now implements `Clone`. ([#554])
 
 [#542]: https://github.com/tower-rs/tower/pull/542
 [#555]: https://github.com/tower-rs/tower/pull/555
 [#557]: https://github.com/tower-rs/tower/pull/557
+[#533]: https://github.com/tower-rs/tower/pull/533
+[#551]: https://github.com/tower-rs/tower/pull/551
+[#554]: https://github.com/tower-rs/tower/pull/554
+[#552]: https://github.com/tower-rs/tower/pull/552
 
 # 0.4.4 (January 20, 2021)